Reorganize Login component
